Fixed CI lint errors by splitting CartContext.tsx into four files.
