--- src/context/CartContext.tsx
@@ -1,65 +1,0 @@
-import { createContext, useState, useContext } from "react";
-import type { CartItem } from "../types/CartItem";
-import type { Product } from "../types/Product";
-import "./CartStatus.css";
-
-//declaring the CartContext type
-type CartContextType = {
-  items: CartItem[];
-  addToCart: (product: Product) => void;
-  removeFromCart: (productId: number) => void;
-  totalItems: number;
-};
-
-export const CartContext = createContext<CartContextType | null>(null);
-
-export const CartProvider = ({ children }: { children?: React.ReactNode }) => {
-  const [items, setItems] = useState<CartItem[]>([]);
-
-  const removeFromCart = (productId: number) => {
-    setItems((prev) => prev.filter((i) => i.product.id !== productId));
-  };
-
-  const totalItems = items.reduce((sum, i) => sum + i.quantity, 0);
-
-  const addToCart = (product: Product) => {
-    setItems((prev) => {
-      const existing = prev.find((i) => i.product.id === product.id);
-      if (existing) {
-        return prev.map((i) =>
-          i.product.id === product.id ? { ...i, quantity: i.quantity + 1 } : i
-        );
-      }
-      return [...prev, { product, quantity: 1 }];
-    });
-  };
-
-  return (
-    <CartContext.Provider
-      value={{ items, addToCart, removeFromCart, totalItems }}>
-      <CartStatus />
-      {children}
-    </CartContext.Provider>
-  );
-};
-
-export const CartStatus = () => {
-  const { totalItems } = useCart();
-  return (
-    <div className="cart-status" role="status" aria-live="polite" aria-label={`${totalItems} in cart`}>
-      {totalItems > 0 && (
-        <span className="cart-badge" aria-hidden="true">
-          {totalItems}
-        </span>
-      )}
-    </div>
-  );
-};
-
-export const useCart = () => {
-  const context = useContext(CartContext);
-  if (context == null) {
-    throw new Error("useCart must be used within a CartProvider");
-  }
-  return context;
-};
