+++ src/context/CartProvider.tsx
@@ -1,0 +1,33 @@
+import { useState } from "react";
+import type { ReactNode } from "react";
+import type { CartItem } from "../types/CartItem";
+import type { Product } from "../types/Product";
+import { CartContext } from "./CartContextDef";
+
+export const CartProvider = ({ children }: { children?: ReactNode }) => {
+  const [items, setItems] = useState<CartItem[]>([]);
+
+  const removeFromCart = (productId: number) => {
+    setItems((prev) => prev.filter((i) => i.product.id !== productId));
+  };
+
+  const totalItems = items.reduce((sum, i) => sum + i.quantity, 0);
+
+  const addToCart = (product: Product) => {
+    setItems((prev) => {
+      const existing = prev.find((i) => i.product.id === product.id);
+      if (existing) {
+        return prev.map((i) =>
+          i.product.id === product.id ? { ...i, quantity: i.quantity + 1 } : i
+        );
+      }
+      return [...prev, { product, quantity: 1 }];
+    });
+  };
+
+  return (
+    <CartContext.Provider value={{ items, addToCart, removeFromCart, totalItems }}>
+      {children}
+    </CartContext.Provider>
+  );
+};
