+++ src/context/useCart.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { CartContext } from "./CartContextDef";
+
+export const useCart = () => {
+  const context = useContext(CartContext);
+  if (context == null) {
+    throw new Error("useCart must be used within a CartProvider");
+  }
+  return context;
+};
